backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mongoengine==0.28.2
pymongo==4.6.0
python-decouple==3.8
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow>=12.0.0
